Restructure project files for better handling multiple events.
